[tests] removed Samples class
Seems better just as a method in the unit test that uses it
