+++ tests/MonoEmbeddinator4000Tests/MonoEmbeddinator4000Tests.csproj
@@ -53,7 +53,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ResourceDesignerTest.cs" />
-    <Compile Include="Helpers\Samples.cs" />
     <Compile Include="JavaNameTest.cs" />
     <Compile Include="XamarinAndroidTest.cs" />
     <Compile Include="Helpers\Approvals.cs" />
